+++ library/stdio/openiob.c
@@ -150,9 +150,11 @@ int __open_iob(const char *filename, const char *mode, int file_descriptor, int 
 	if (FLAG_IS_SET(open_mode, O_WRONLY) || FLAG_IS_SET(open_mode, O_RDWR))
 		SET_FLAG(file_flags, IOBF_WRITE);
 
-	__initialize_iob(file, __iob_hook_entry,
+	__initialize_iob(file,
+                     __iob_hook_entry,
 					 buffer,
-					 aligned_buffer, BUFSIZ,
+					 aligned_buffer,
+                     (int64_t) BUFSIZ,
 					 file_descriptor,
 					 slot_number,
 					 file_flags,
